refactor SWR block updates to custom hook
