+++ src/data/Allowances.ts
@@ -1,11 +1,9 @@
-import { useEffect, useRef } from 'react'
 import { Contract } from '@ethersproject/contracts'
 import { Token, TokenAmount } from '@uniswap/sdk'
 import useSWR from 'swr'
 
-import { SWRKeys } from '.'
+import { SWRKeys, useKeepSWRDataLiveAsBlocksArrive } from '.'
 import { useTokenContract } from '../hooks'
-import { useBlockNumber } from '../state/application/hooks'
 
 function getTokenAllowance(contract: Contract, token: Token): (owner: string, spender: string) => Promise<TokenAmount> {
   return async (owner: string, spender: string): Promise<TokenAmount> =>
@@ -22,16 +20,7 @@ export function useTokenAllowance(token?: Token, owner?: string, spender?: strin
     shouldFetch ? [owner, spender, token.address, token.chainId, SWRKeys.Allowances] : null,
     getTokenAllowance(contract, token)
   )
-
-  // fetch data again every time there's a new block
-  const mutateRef = useRef(mutate)
-  useEffect(() => {
-    mutateRef.current = mutate
-  })
-  const blockNumber = useBlockNumber()
-  useEffect(() => {
-    mutateRef.current()
-  }, [blockNumber])
+  useKeepSWRDataLiveAsBlocksArrive(mutate)
 
   return data
 }
